removed Notes
Removed Notes tab in the export file. 

p.s. This sheet of the export file was populated from the Notes tab of the Management DataStore (backend sheet of the field management data).  Information contained in the notes now is populated in corresponding site edit sheets at SS. We will need to generate report in SS and use that instead.